fnct: uniform 'private/public' ones
And words order: xx_yy_one and xx_yy_all instead of xx_one_yy and
xx_all_yy, easier for the grep.

Prefixing with _ to show which functions should not be used directly if
we extend this script for tempory extra steps.

(also to avoid replacing binary/scripts, e.g. tap())

Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>