timeout: switch to 55 minutes
For the moment, the limit on Cirrus is set to 60 minutes. We can
increase it if needed but should be enough so far.

By increasing the number of tests, we also increase the time it takes to
run all tests. We will need to work on that but for the moment
(hopefully), increase the global timeout for expect (not taking into
account the compilation time).

Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>