packetdrill: properly increase the tolerance
The hacky 'sed' command we had for the 'normal' mode was not working for
some tests because we were doing 'sed s/=1/=2' while 'add_addr' tests
have a tolerance of 200000 and not 100000 while all the others.

Here we properly get the value, multiply it and then replace it in the
packetdrill files.

Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>