timeout: increase to 75min
We have more and more tests and the public CI on Cirrus is very slow
when using the debug mode.

We already reached this timeout a few times, it is time to increase it.

For the moment, the global timeout is set at 90min. Here we restrict the
tests to run for max 75min.

Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>