kconfig: move debug config to where it is used
On one side, it avoids confusions where we could see at the beginning of
the log with 'set -x' enabled that some debug kconfig are added but they
are only added in an array that is not necessarily used later.

On the other side, we will soon need to do different actions in the
kconfig depending on the mode.

Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>