git: Avoid invoking 'git' as root
Latest versions of git will fail if the owner of the repo (in terms
of file permissions) is different from the current user.

prepare() in entrypoint.sh was running 'git rev-parse' as root, which
triggers a fatal error:

Prepare the environment
+ is_ci
+ '[' '' = true ']'
++ git rev-parse --short HEAD
fatal: unsafe repository ('/home/example-user/linux' is owned by someone else)

Instead, run 'git rev-parse' in run.sh (before changing uid) and pass in
the value using the docker environment. This works ok when using run.sh
or scripts that call it, but requires more command line boilerplate when
directly using the 'docker run' command as described in the README.

An alternative would be to extract the git rev from .git/HEAD and
.git/refs/heads/<branchname-from-.git/HEAD>, but that's a fragile
way to go about it.

Signed-off-by: Mat Martineau <mathew.j.martineau@linux.intel.com>