git: avoid 'unsafe repository' error
We don't modify the kernel git repo but we need to get info from it: rev
and nearest tag.

With the recent update to fix CVE-2022-24765, these 'git' commands fail
with 'unsafe repository' error.

To fix this, the global config in the docker image needs to be updated
to mark this directory as "safe".

Closes: #3
Reported-by: Mat Martineau <mathew.j.martineau@linux.intel.com>
Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>
(cherry picked from commit 1b0867d601888658e2f9a7a41b1947f449fd89db)
Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>