docker: pktdrill: also install run_all.py
This is mainly a workaround to force a rebuild of packetdrill with the
latest version because a 'make clean' is needed after having sync with
upstream.

But it is also nice to have run_all.py available, even if we need to
launch it from /opt/packetdrill/gtests/net dir.

Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>