virtme: support kernel 6.2
Since the kernel commit f65a486821cf ("kbuild: change module.order to
list *.o instead of *.ko") that is now in Linus tree in preparation of
the future v6.2, module.order file lists .o files instead of .ko ones.

virtme-prep-kdir-mods reads module.order file but it needs to create
symlinks for the .ko files, not the .o ones.

We then need a patched version of virtme to be able to start the kernel
and load required modules.

Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>