kconfig: force panic in case of soft lockup/hung tasks
For the moment, issues are printed and the job continue doing nothing
until the global timeout hit.

In the last case we had, it is not even possible to stop the task with
'Ctrl+C' + 'poweroff', so we don't have the "post" tasks analysing the
stack trace, saving stuff, etc.

Example:

  https://cirrus-ci.com/task/5999689074475008

Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>