tests: support subtests with embedded TAP results
Some selftests like the ones from TC already print embedded TAP 13
results in order to track the different subtests and not just a global
result for a bunch of tests.

The idea here is simply to extract such results in a dedicated file for
the CI to easily track all the subtests if available. If not, the
behaviour is not supposed to be modified so we can use this modification
with older versions.

This patch also sets SELFTESTS_MPTCP_LIB_NO_TAP=1 when it is not the CI
which is using this script in order not to print the detailed results
formatted in TAP 13 output as the results are already presented in a
human-friendly version before.

Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>