summary: print only failed subtests
With the new support of the selftests, in case of errors, all subtests
that have been executed are printed and all the ones for the selftests
that have failed are printed again.

We don't need that in the summary, we want to focus on what didn't work.

We can easily filter them by looking if the result is followed by
'test:' or ' - <selftest>'.

Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>