timeout: new dedicated boot timeout
We recently reduced the global timeout because we had a few issues in
the past where the console would only arrive after one hour.

Instead, we can use a different timeout for the boot process (3 minutes)
and increase the global one to avoid having not enough time to run the
tests after the clone + compilation, etc.

Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>