timeout: extra timeout of 60s to get the console
The CI recently complained that it was not able to get the console even
if the message 'virtme-init: console is ttyS0' has been seen.

Once this message has been seen, we can try to get the console by
pressing "Enter" every second. If after 1 min, we still didn't get it,
we can exit with an error instead of waiting for 2h.

Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>