bpftest: test_progs are compiled in the build dir
It looks a bit strange to have all the BPF selftests compiled files in
the root of the build dir. I tried to set:

  O=<BUILD DIR>/tools/testing/selftests/bpf'

... but I ran into different issues, mainly compiling bpf_testmod we
don't need. KDIR=<BUILD DIR> should have helped I think, but I guess I
missed something.

It would be good to keep the root of the build dir for the kernel, and
place built files from BPF selftests elsewhere. This can be done later.

Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>