virtme: switch to virtme-ng
'virtme' is no longer maintained, we had to fork it to fix issues [1].

The new 'virtme-ng' brings fixes, new features, and performances
improvement, especially for the disk IO [2].

Note that three workarounds are currently needed:
 - The new init in rust cannot be used [3].
 - The .config needs to be removed [4].
 - The whole rootfs is mounted in rw instead of only the workdir to be
   able to used virtiofs (faster) instead of 9p. But that's OK, our
   scripts are launched from a docker where only the workdir is mounted.
   Maybe even easier like that.

There are new dependences, but it is still light. 'udev' has been added
to avoid a warning, but I don't think we need it.

Link: https://github.com/matttbe/virtme [1]
Link: https://lwn.net/Articles/951313/ [2]
Link: https://github.com/arighi/virtme-ng/issues/90 [3]
Link: https://github.com/arighi/virtme-ng/pulls/91 [4]
Closes: https://github.com/multipath-tcp/mptcp_net-next/issues/472
Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>