gcov: add GCOV support
When launched with INPUT_GCOV=1, the kernel will be compiled with GCOV
support. At the end of the tests, lcov will be used to extract the
required info. In the analysis phase, an HTML report will be generated,
only for net/mptcp.

Note that -fprofile-update=atomic is used to avoid corrupted counters,
but still, some can be negative, and '--rc geninfo_unexecuted_blocks=1'
is then needed.

The HTML report can be directly displayed, and the .lcov file can be
used to track the evolution.

Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>