expect: timeout the shutdown part
Sometimes, the shutdown part gets stuck: in this case, and after a
couple of minutes, the kernel detects that (hung_task), panic, but
that's not enough because then the VM got stuck with 'KVM: entry failed,
hardware error' message and a stack trace.

We can set a timeout of 1 minute when shutting down the VM, and exit
after that if the process didn't finish on time.

Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>