build: dedicated dir for 'debug' mode
To avoid issues when switching from one to another, plus to keep minimal
cache, the debug mode now has its own build and ccache dirs.

The 'all' modes now simply relaunch the script with the different modes.
That's also easier to deal with that to avoid appending values twice on
the same variable.

A summary is printed at the end, presenting all the results.

Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>