timeout: trigger sysrq to help debug
This cycle, some jobs got blocked when running the tests, and the CI
didn't show any useful info in case of a timeout.

The idea here is to have a simple script running in a background,
waiting for the timeout to trigger some sysrq in order to get some
useful debug info.

Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>