docker: build: pull the image
docker build will pull the 'FROM' image layers if it is missing, but it
will not tag it. It means that the next builds will query the registry
to check if an updated version is available.

That's certainly good for production image, but not for dev ones, where
we want a quick build, no latency when rebuilding and there is nothing
to build, and no need to pull the latest version all the time (which
will cause a rebuild of everything, which can be slow...).

Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>