expect: workaround to get serial output in sync
It looks like expect will only display the output when it processes it,
i.e. when using 'expect*' commands.

We only need the serial at boot and exit time, in between stdout can be
consumed directly thanks to the expect_background command.

This command has no timeout, but it blocks other 'expect' commands: it
is then needed to stop it when there is a need to read the serial
output: this can be done by calling expect_background without any
arguments.

Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>