timeout: get more info via GDB
The VM running in the CI occasionally freezes, and when it happens, it
is not possible to interact with it via the console nor the serial
any more.

Hopefully GDB will help us understanding what's wrong.

Maybe there are other GDB commands to pass, we can always modify the
list later.

Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>