kconfig: no need to force RCU_EXPERT
This option is now enabled by default in the debug kconfig since v6.13.
It is needed to enable PROVE_RCU_LIST.

It is better not to force it, because a few false positive issues have
been fixed only in v6.13. If we force it, we will get them when
validating older kernel versions.

Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>