results dir: replace container ID by hostname
This container/docker ID is unique, but not clear and not predictable.
Instead, we can prepend the given hostname and CID, clearer and unique
enough.

While at it, also use the mode before the hostname, so all builds from
the same container will be under the same directory, i.e.

  <sha>/<mode>/<host>, e.g. 123456/normal/{client,server}

And not in different parts, i.e.

  <sha>/<host>/<mode>, e.g. 123456/{client,server}/normal

Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>