bridge: setup netem rules on bridge's ifaces
Instead of on the bridge iface itself, which will not do anything in our
case.

So these rules should be set after having started all VMs, and the
interfaces have been created after the DHCP request. Waiting for the
prompt is then enough.

The logic has then been moved to the Python code which is cleaner
anyway.

Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>