ci: timeout: switch to 2h
Recently, the timeout was reached again [1].

Switch to 2h as it is the current hard timeout anyway. The script will
stop before to have some time to print extra debug about the current
state, see the comment around VIRTME_CI_ESTIMATED_EXTRA_TIME.

We might want to execute more tasks in parallel at some points to save
the global time if we start going over 2h for the tests. Or accelerate
the tests somehow.

Link: https://github.com/multipath-tcp/mptcp_net-next/actions/runs/17315894916/job/49158561749 [1]
Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>